--- app/src/main/java/com/techcourse/HandlerAdapterRegistry.java
@@ -1,4 +1,0 @@
-package com.techcourse;
-
-public class HandlerAdapterRegistry {
-}
