+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/HandlerMappingRegistry.java
@@ -1,0 +1,4 @@
+package com.interface21.webmvc.servlet.mvc;
+
+public class HandlerMappingRegistry {
+}
